Updated product page Made brand relationship to product Changed the type of volume in product tables Updated product factory

+++ database/migrations/2018_12_08_015426_add_volume_to_products_table.php
@@ -14,7 +14,7 @@ class AddVolumeToProductsTable extends Migration
     public function up()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->decimal('volume', 8, 2)->nullable()->after('price');
+            $table->integer('volume')->nullable()->after('price');
         });
     }
